+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- Add git revision to "sr version" output
+
 ## v0.12.0 (2023-01-10)
 
 - Rename `run_embedded` step property to `run-embedded`. `run_embedded` is still accepted as an alias in `sr.yaml` files. [#2](https://github.com/insilica/rs-srvc/issues/2)